backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Werkzeug==3.1.3
openpyxl==3.1.5
reportlab==4.4.0
Pillow==11.1.0
psycopg2-binary==2.9.9
pandas==2.2.3
Flask-Limiter==3.10.1
Flask-Mail==0.10.0
google-auth==2.38.0
requests==2.32.3
